requirements/base.txt
django==3.1.5
django-markdownx==3.0.1
